+++ project/City.h
@@ -8,12 +8,12 @@ private:
     int id;
     std::string code;
     std::string name;
-    double demand;
+    int demand;
     unsigned population;
 public:
     City(int id, std::string code, std::string name, double demand, int population): id(id), code(std::move(code)), name(std::move(name)), demand(std::move(demand)), population(population) {};
     [[nodiscard]] std::string getName() const;
-    [[nodiscard]] double getDemand() const;
+    [[nodiscard]] int getDemand() const;
     [[nodiscard]] unsigned getPopulation() const;
     int getID() const;
     std::string getCode();
